REF: Gather the experiments shared helpers into _base and _io
